refactor delegation check logic and extract hook for reusability
